requirements.txt
PyVirtualDisplay==2.1
selenium-firefox==1.0.33